fix: update type handling in various components to use size_t for index variables and ensure proper casting in expressions

+++ src/IR/IRGenerator.h
@@ -138,7 +138,7 @@ public:
 
   std::unique_ptr<IRParser> &getIRParserPtr();
   void setModuleCount(int count);
-  int32_t hasErrors() const;
+  int8_t hasErrors() const;
 
 private:
   llvm::LLVMContext *TheContext;
